Material design refresh - phase 1
