+++ .jshintrc
@@ -2,9 +2,7 @@
   "predef": [
     "document",
     "window",
-    "-Promise",
-    "nv",
-    "d3"
+    "-Promise"
   ],
   "browser": true,
   "boss": true,
