+++ app/components/druid-navbar.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+	tagName: 'nav',
+	classNames: ['navbar', 'druid-navbar', 'navbar-inverse', 'navbar-static-top'],
+	role: 'navigation',
+	attributeBindings: ['role']
+});
