+++ app/routes/datasource/index.js
@@ -17,7 +17,7 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
-  model: function () {
-    return this.store.findAll('datasource');
+  redirect: function () {
+    return this.transitionTo('datasources.index');
   }
 });
