+++ app/styles/app.scss
@@ -1,36 +1,13 @@
-/*
- * Copyright (c) 2015 Yahoo Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+@import 'materialize';
 
-.ember-date-time-picker {
-	margin-bottom: 10px;
-	.input-group-addon {
-		width: 20px;
-	}
+/* FIXME: Workaround for https://github.com/Dogfalo/materialize/issues/1079 */
+@each $mdi-icon-name, $mdi-icon-value in $mdi-list-icons {
+  .#{$mdi-prefix}#{$mdi-icon-name}:before {
+    content: "\""+ $mdi-icon-value +"\"";
+  }
 }
 
-.top-k-list {
-	.tooltip-inner {
-		max-width: 300px;
-		word-wrap: break-word;
-	}
-
-	.value-cell {
-		color: #bbb;
-		.suffix {
-			color: #999;
-		}
-	}
+.avatar-circle {
+	width: 40px;
+	height: 40px;
 }
