+++ app/templates/application.hbs
@@ -1,12 +1,5 @@
-<nav class="navbar navbar-inverse navbar-static-top" role="navigation">
-	<div class="container">
-		<div class="navbar-header">
-			{{#link-to 'index' class='navbar-brand'}}
-			   <img src="img/druid_nav.png" height="22">
-			{{/link-to}}
-		</div>
-	</div>
-</nav>
+{{druid-navbar}}
+
 <div id="content" class="container">
 	{{outlet}}
 </div>
