+++ app/templates/components/druid-navbar.hbs
@@ -1,0 +1,21 @@
+<div class="nav-wrapper">
+	<div class="container">
+		{{#link-to 'index' class='brand-logo'}}
+			<img src="img/druid_nav.png" height="22">
+		{{/link-to}}
+		<a href="#" data-activates="mobile-demo" class="button-collapse">
+			<i class="mdi-navigation-menu"></i>
+		</a>
+		<ul class="right hide-on-med-and-down">
+			{{#link-to 'datasource.index' active='active' tagName='li'}}
+				<a> Datasources</a>
+			{{/link-to}}
+      	</ul>
+      	<ul class="side-nav" id="mobile-demo">
+	        <li><a href="sass.html">Sass</a></li>
+	        <li><a href="components.html">Components</a></li>
+	        <li><a href="javascript.html">Javascript</a></li>
+	        <li><a href="mobile.html">Mobile</a></li>
+      	</ul>
+    </div>
+</div>
