--- app/templates/datasource/index.hbs
@@ -1,9 +1,0 @@
-
-<div class="row text-center">
-  <h1>Datasources</h1>
-  <div class="btn-group-vertical col-md-4 col-md-offset-4">
-    {{#each content as |item|}}
-      {{link-to item.id 'datasource.show' item.id class="btn btn-default"}}
-    {{/each}}
-  </div>
-</div>
