+++ app/templates/datasource/show.hbs
@@ -1,84 +1,30 @@
 <div class="row">
-  <div class="col-md-12">
-    <h1>Datasource - {{model.id}}</h1>
+  <div class="col s12">
+    <h4>Datasource - {{text-titlecase model.id}}</h4>
   </div>
 </div>
-
 <div class="row">
-    <div class="col-sm-3">
-      {{dimensionFilters.region}}
+	<div class="col m3 s12">
+		{{materialize-date-input label='Start Date' value=startDateStr}}
+	</div>
+	<div class="col m3 s12">
+		{{materialize-date-input label='End Date' value=endDateStr}}
+	</div>
+	<div class="col m3 s12">
+		{{materialize-select
+			label="Layout"
+			content=allLayoutModes
+			value=layoutMode
+			optionValuePath="content.id"
+			optionLabelPath="content.label"}}
+	</div>
+	<div class="col m3 s12">
+		{{materialize-select
+			label="Time Granularity"
+			content=allTimeGranularities
+			value=timeGranularity
+			optionValuePath="content.id"
+			optionLabelPath="content.label"}}
+	</div>
 
-        <div class="form-group">
-            <label>Time Granularity</label>
-
-            <div class="form-component">
-              {{view 'e-select'
-              content=allTimeGranularities
-              value=timeGranularity}}
-            </div>
-        </div>
-    </div>
-    <div class="col-sm-3">
-      {{date-time-picker content=startDate label="From"}}
-    </div>
-    <div class="col-sm-3">
-      {{date-time-picker content=endDate label="To"}}
-    </div>
-    <div class="col-sm-3">
-        <div class="form-component">
-            <label>Layout</label>
-          {{view 'e-select'
-          content=layoutModes
-          value=dashLayout}}
-        </div>
-    </div>
-</div>
-<div class="row">
-    <div class="col-sm-3">
-        <div class="form-group">
-            <label>Metric</label>
-
-            <div class="form-component">
-              {{view 'e-select'
-              content=availableMetrics
-              value=metric}}
-            </div>
-        </div>
-    </div>
-    <div class="col-sm-6">
-        <table class="table">
-            <thead>
-            <tr>
-                <th colspan="2">Filters</th>
-            </tr>
-            </thead>
-            <tbody>
-            {{#each filters as |dat|}}
-                <tr {{action 'filterClicked' dat.dim}}>
-                    <td>
-                      {{dat.dim}}
-                    </td>
-                    <td class='text-right'>
-                      {{dat.val}}
-                    </td>
-                </tr>
-            {{/each}}
-            </tbody>
-        </table>
-    </div>
-</div>
-<div class="row">
-    <div {{bind-attr class="isColumnLayout:col-sm-5:col-sm-12 isColumnLayout:pull-right"}}>
-
-      {{line-graph-stack
-      series=timeseriesData.content.data
-      dashLayout=dashLayout}}
-    </div>
-    <div {{bind-attr class="isColumnLayout:col-sm-7:col-sm-12"}}>
-      {{#each topKData as |tkd|}}
-          <div {{bind-attr class="isColumnLayout:col-sm-6:col-sm-3"}}>
-            {{top-k-list content=tkd action="topKRowClicked"}}
-          </div>
-      {{/each}}
-    </div>
 </div>
