+++ app/templates/datasource/show_old.hbs
@@ -1,0 +1,85 @@
+<div class="row">
+  <div class="col s12">
+    <h4>Datasource - {{model.id}}</h4>
+  </div>
+</div>
+
+<div class="row">
+    <div class="col s3">
+      {{dimensionFilters.region}}
+        <div class="form-group">
+            <label>Time Granularity</label>
+
+            <div class="form-component">
+              {{materialize-select
+              content=allTimeGranularities
+              value=timeGranularity
+              optionValuePath='content'
+              optionLabelPath='content'}}
+            </div>
+        </div>
+    </div>
+    <div class="col s3">
+      {{!-- {{date-time-picker content=startDate label="From"}} --}}
+    </div>
+    <div class="col s3">
+      {{!-- {{date-time-picker content=endDate label="To"}} --}}
+    </div>
+    <div class="col s3">
+        <div class="form-component">
+            <label>Layout</label>
+          {{materialize-select
+          content=layoutModes
+          value=dashLayout}}
+        </div>
+    </div>
+</div>
+<div class="row">
+    <div class="col s3">
+        <div class="form-group">
+            <label>Metric</label>
+
+            <div class="form-component">
+              {{view 'e-select'
+              content=availableMetrics
+              value=metric}}
+            </div>
+        </div>
+    </div>
+    <div class="col s6">
+        <table class="table">
+            <thead>
+            <tr>
+                <th colspan="2">Filters</th>
+            </tr>
+            </thead>
+            <tbody>
+            {{#each filters as |dat|}}
+                <tr {{action 'filterClicked' dat.dim}}>
+                    <td>
+                      {{dat.dim}}
+                    </td>
+                    <td class='text-right'>
+                      {{dat.val}}
+                    </td>
+                </tr>
+            {{/each}}
+            </tbody>
+        </table>
+    </div>
+</div>
+<div class="row">
+    <div {{bind-attr class=":col isColumnLayout:s5:s12 isColumnLayout:pull-right"}}>
+
+      {{line-graph-stack
+      series=timeseriesData.content.data
+      dashLayout=dashLayout}}
+    </div>
+    <div {{bind-attr class=":col isColumnLayout:s7:s12"}}>
+      {{#each topKData as |tkd|}}
+          <div {{bind-attr class=":col isColumnLayout:s6:s3"}}>
+            {{top-k-list content=tkd action="topKRowClicked"}}
+          </div>
+      {{/each}}
+    </div>
+</div>
