+++ app/utils/moment-as-ms.js
@@ -1,0 +1,13 @@
+import moment from 'moment';
+
+export default function momentAsMs(prop) {
+  return Ember.computed(prop, {
+  	get() {
+  		return this.get(prop).valueOf();
+  	},
+  	set(key, val) {
+     	this.set(prop, moment(val));
+      	return val;
+  	}
+  });
+}
