+++ bower.json
@@ -2,23 +2,23 @@
   "name": "druid-ui",
   "dependencies": {
     "jquery": "^1.11.1",
-    "ember": "1.11.3",
+    "ember": "1.12.0-beta.2",
     "ember-data": "1.0.0-beta.16.1",
-    "ember-resolver": "~0.1.14",
+    "ember-resolver": "~0.1.15",
     "loader.js": "ember-cli/loader.js#3.2.0",
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
-    "ember-load-initializers": "ember-cli/ember-load-initializers#0.0.2",
-    "ember-qunit": "0.2.8",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
+    "ember-qunit": "0.3.1",
     "ember-qunit-notifications": "0.0.7",
     "qunit": "~1.17.1",
-    "bootstrap": "~3.3.2",
-    "ember-addons.bs_for_ember": "~0.7.0",
-    "bootstrap-datetimepicker": "git@github.com:smalot/bootstrap-datetimepicker.git#2.3.1",
     "ember-cli-moment-shim": "~0.0.3",
-    "bootstrap-select": "~1.6.3",
-    "nvd3": "~1.7.1",
     "lodash": "~3.2.0",
-    "ember-data.model-fragments": "0.3.1"
+    "ember-data.model-fragments": "0.3.1",
+    "materialize": "~0.96.1",
+    "pretender": "0.1.0"
+  },
+  "resolutions": {
+    "jquery": "^1.11.1"
   }
 }
