+++ package.json
@@ -19,21 +19,24 @@
   "license": "MIT",
   "devDependencies": {
     "broccoli-asset-rev": "^2.0.2",
-    "ember-cli": "0.2.1",
+    "ember-cli": "0.2.3",
     "ember-cli-app-version": "0.3.3",
-    "ember-cli-babel": "^4.0.0",
+    "ember-cli-babel": "^5.0.0",
     "ember-cli-content-security-policy": "0.4.0",
     "ember-cli-dependency-checker": "0.0.8",
     "ember-cli-htmlbars": "0.7.4",
     "ember-cli-ic-ajax": "0.1.1",
     "ember-cli-inject-live-reload": "^1.3.0",
-    "ember-cli-qunit": "0.3.9",
-    "ember-cli-sass": "3.2.1",
+    "ember-cli-materialize": "https://github.com/truenorth/ember-cli-materialize.git#all",
+    "ember-cli-pretender": "0.3.1",
+    "ember-cli-qunit": "0.3.10",
+    "ember-cli-sass": "4.0.0-beta.5",
     "ember-cli-uglify": "1.0.1",
     "ember-data": "1.0.0-beta.16.1",
     "ember-data-model-fragments": "^0.3.1",
     "ember-export-application-global": "^1.0.2",
     "ember-moment": "1.0.0",
+    "ember-new-computed": "1.0.0",
     "ember-try": "0.0.4"
   }
 }
