+++ tests/unit/helpers/text-titlecase-test.js
@@ -1,0 +1,12 @@
+import {
+  textTitlecase
+} from '../../../helpers/text-titlecase';
+import { module, test } from 'qunit';
+
+module('TextTitlecaseHelper');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var result = textTitlecase(42);
+  assert.ok(result);
+});
