+++ tests/unit/utils/moment-as-ms-test.js
@@ -1,0 +1,10 @@
+import momentAsMs from '../../../utils/moment-as-ms';
+import { module, test } from 'qunit';
+
+module('momentAsMs');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var result = momentAsMs();
+  assert.ok(result);
+});
