+++ tests/unit/utils/moment-as-str-test.js
@@ -1,0 +1,10 @@
+import momentAsStr from '../../../utils/moment-as-str';
+import { module, test } from 'qunit';
+
+module('momentAsStr');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var result = momentAsStr();
+  assert.ok(result);
+});
